+++ src/tools/getLatestPipelineStatus/inputSchema.ts
@@ -29,6 +29,8 @@ export const getLatestPipelineStatusInputSchema = z.object({
     .describe(
       'The URL of the CircleCI project. Can be any of these formats:\n' +
         '- Project URL with branch: https://app.circleci.com/pipelines/gh/organization/project?branch=feature-branch\n' +
+        '- Legacy Pipeline URL: https://circleci.com/gh/organization/project/123\n' +
+        '- Legacy Pipeline URL with branch: https://circleci.com/gh/organization/project/123?branch=feature-branch\n' +
         '- Pipeline URL: https://app.circleci.com/pipelines/gh/organization/project/123\n' +
         '- Workflow URL: https://app.circleci.com/pipelines/gh/organization/project/123/workflows/abc-def\n' +
         '- Job URL: https://app.circleci.com/pipelines/gh/organization/project/123/workflows/abc-def/jobs/xyz',